+++ src/lib.rs
@@ -3,3 +3,4 @@ pub mod web;
 pub mod util;
 pub mod backup;
 pub mod cli;
+pub mod protocol;