+++ src/util/mod.rs
@@ -2,3 +2,4 @@ pub mod progress_bar;
 pub mod java_util;
 pub mod runner_util;
 pub mod logger;
+pub mod paper_api;